messaging: add ping/pong helpers and migrate call sites to NDR-aware API
Add messaging_send_ping(), messaging_send_all_ping(),
messaging_register_pong(), and messaging_read_send_pong() helpers to
source3/lib/messages.c and declare them in source3/include/messages.h.

Add imessaging_send_ping() helper to source4/lib/messaging/messaging_send.c
and declare it in source4/lib/messaging/messaging.h.

Migrate all remaining raw MSG_PING/MSG_PONG call sites to use the new
NDR-aware helpers, ensuring the ping handler's NDR pull/push is
correctly exercised:

- source3/torture/msgtest.c
- source3/torture/test_messaging_read.c
- source3/torture/test_messaging_send_all.c
- source3/smbd/notifyd/tests.c
- source4/lib/messaging/tests/messaging.c
- testsuite/unittests/test_background_send.c (cast fix)

+++ testsuite/unittests/test_background_send.c
@@ -34,7 +34,7 @@ static void test_background_send(void)
 	struct tevent_context *ev = NULL;
 	struct messaging_context *msg_ctx = NULL;
 	struct tevent_req *req = NULL;
-	uint32_t ping_msg = MSG_PING;
+	uint32_t ping_msg = (uint32_t)MSG_PING;
 
 	ev = tevent_context_init(frame);
 	assert(ev != NULL);
